Refactor some mixins
