+++ src/mixins/sass/ids-dirty-tracker-mixin.scss
@@ -2,7 +2,7 @@
 
 .icon-dirty {
   @include absolute();
-  @include m-4();
+  @include m-2();
   @include text-alert-caution();
 
   top: 0;
